+++ app/components/workbench/FileBreadcrumb.tsx
@@ -1,14 +1,10 @@
-import * as DropdownMenu from '@radix-ui/react-dropdown-menu';
-import { AnimatePresence, motion, type Variants } from 'framer-motion';
 import { memo, useEffect, useRef, useState } from 'react';
-import FileTree from './FileTree';
+import * as DropdownMenu from '@radix-ui/react-dropdown-menu';
+import { BreadcrumbSegment } from './BreadcrumbSegment';
+import { BreadcrumbDropdown } from './BreadcrumbDropdown';
 import type { FileMap } from '~/lib/stores/files';
-import { classNames } from '~/utils/classNames';
-import { WORK_DIR } from '~/utils/constants';
-import { cubicEasingFn } from '~/utils/easings';
+import { WORK_DIR_REGEX } from '~/utils/constants';
 import { renderLogger } from '~/utils/logger';
-
-const WORK_DIR_REGEX = new RegExp(`^${WORK_DIR.split('/').slice(0, -1).join('/').replaceAll('/', '\\/')}/`);
 
 interface FileBreadcrumbProps {
   files?: FileMap;
@@ -16,30 +12,10 @@ interface FileBreadcrumbProps {
   onFileSelect?: (filePath: string) => void;
 }
 
-const contextMenuVariants = {
-  open: {
-    y: 0,
-    opacity: 1,
-    transition: {
-      duration: 0.15,
-      ease: cubicEasingFn,
-    },
-  },
-  close: {
-    y: 6,
-    opacity: 0,
-    transition: {
-      duration: 0.15,
-      ease: cubicEasingFn,
-    },
-  },
-} satisfies Variants;
-
 export const FileBreadcrumb = memo<FileBreadcrumbProps>(({ files, pathSegments = [], onFileSelect }) => {
   renderLogger.trace('FileBreadcrumb');
 
   const [activeIndex, setActiveIndex] = useState<number | null>(null);
-
   const contextMenuRef = useRef<HTMLDivElement | null>(null);
   const segmentRefs = useRef<(HTMLSpanElement | null)[]>([]);
 
@@ -59,10 +35,7 @@ export const FileBreadcrumb = memo<FileBreadcrumbProps>(({ files, pathSegments =
     };
 
     document.addEventListener('mousedown', handleOutsideClick);
-
-    return () => {
-      document.removeEventListener('mousedown', handleOutsideClick);
-    };
+    return () => document.removeEventListener('mousedown', handleOutsideClick);
   }, [activeIndex]);
 
   if (files === undefined || pathSegments.length === 0) {
@@ -73,7 +46,6 @@ export const FileBreadcrumb = memo<FileBreadcrumbProps>(({ files, pathSegments =
     <div className="flex">
       {pathSegments.map((segment, index) => {
         const isLast = index === pathSegments.length - 1;
-
         const path = pathSegments.slice(0, index).join('/');
 
         if (!WORK_DIR_REGEX.test(path)) {
@@ -85,63 +57,29 @@ export const FileBreadcrumb = memo<FileBreadcrumbProps>(({ files, pathSegments =
         return (
           <div key={index} className="relative flex items-center">
             <DropdownMenu.Root open={isActive} modal={false}>
-              <DropdownMenu.Trigger asChild>
-                <span
-                  ref={(ref) => {
-                    segmentRefs.current[index] = ref;
-                    return undefined;
-                  }}
-                  className={classNames('flex items-center gap-1.5 cursor-pointer shrink-0', {
-                    'text-bolt-elements-textTertiary hover:text-bolt-elements-textPrimary': !isActive,
-                    'text-bolt-elements-textPrimary underline': isActive,
-                    'pr-4': isLast,
-                  })}
-                  onClick={() => handleSegmentClick(index)}
-                >
-                  {isLast && <div className="i-ph:file-duotone" />}
-                  {segment}
-                </span>
-              </DropdownMenu.Trigger>
-              {index > 0 && !isLast && <span className="i-ph:caret-right inline-block mx-1" />}
-              <AnimatePresence>
-                {isActive && (
-                  <DropdownMenu.Portal>
-                    <DropdownMenu.Content
-                      className="z-file-tree-breadcrumb"
-                      asChild
-                      align="start"
-                      side="bottom"
-                      avoidCollisions={false}
-                    >
-                      <motion.div
-                        ref={contextMenuRef}
-                        initial="close"
-                        animate="open"
-                        exit="close"
-                        variants={contextMenuVariants}
-                      >
-                        <div className="rounded-lg overflow-hidden">
-                          <div className="max-h-[50vh] min-w-[300px] overflow-scroll bg-bolt-elements-background-depth-1 border border-bolt-elements-borderColor shadow-sm rounded-lg">
-                            <FileTree
-                              files={files}
-                              hideRoot
-                              rootFolder={path}
-                              collapsed
-                              allowFolderSelection
-                              selectedFile={`${path}/${segment}`}
-                              onFileSelect={(filePath) => {
-                                setActiveIndex(null);
-                                onFileSelect?.(filePath);
-                              }}
-                            />
-                          </div>
-                        </div>
-                        <DropdownMenu.Arrow className="fill-bolt-elements-borderColor" />
-                      </motion.div>
-                    </DropdownMenu.Content>
-                  </DropdownMenu.Portal>
-                )}
-              </AnimatePresence>
+              <BreadcrumbSegment
+                segment={segment}
+                isLast={isLast}
+                isActive={isActive}
+                onClick={() => handleSegmentClick(index)}
+                segmentRef={(ref) => {
+                  segmentRefs.current[index] = ref;
+                  return undefined;
+                }}
+              />
+              <BreadcrumbDropdown
+                isActive={isActive}
+                path={path}
+                segment={segment}
+                files={files}
+                contextMenuRef={(ref) => {
+                  if (contextMenuRef.current !== ref) {
+                    contextMenuRef.current = ref;
+                  }
+                }}
+                onFileSelect={onFileSelect}
+                onClose={() => setActiveIndex(null)}
+              />
             </DropdownMenu.Root>
           </div>
         );
