+++ app/lib/stores/artifacts/ArtifactStore.ts
@@ -1,0 +1,89 @@
+import { map, type MapStore } from 'nanostores';
+import { ActionRunner } from '~/lib/runtime/action-runner';
+import type { ActionCallbackData, ArtifactCallbackData } from '~/lib/runtime/message-parser';
+import { webcontainer } from '~/lib/webcontainer';
+import { unreachable } from '~/utils/unreachable';
+
+export interface ArtifactState {
+  id: string;
+  title: string;
+  closed: boolean;
+  runner: ActionRunner;
+}
+
+export type ArtifactUpdateState = Pick<ArtifactState, 'title' | 'closed'>;
+type Artifacts = MapStore<Record<string, ArtifactState>>;
+
+export class ArtifactStore {
+  artifacts: Artifacts = import.meta.hot?.data.artifacts ?? map({});
+  artifactIdList: string[] = [];
+
+  constructor() {
+    if (import.meta.hot) {
+      import.meta.hot.data.artifacts = this.artifacts;
+    }
+  }
+
+  get firstArtifact(): ArtifactState | undefined {
+    return this.#getArtifact(this.artifactIdList[0]);
+  }
+
+  addArtifact({ messageId, title, id }: ArtifactCallbackData) {
+    const artifact = this.#getArtifact(messageId);
+
+    if (artifact) {
+      return;
+    }
+
+    if (!this.artifactIdList.includes(messageId)) {
+      this.artifactIdList.push(messageId);
+    }
+
+    this.artifacts.setKey(messageId, {
+      id,
+      title,
+      closed: false,
+      runner: new ActionRunner(webcontainer),
+    });
+  }
+
+  updateArtifact({ messageId }: ArtifactCallbackData, state: Partial<ArtifactUpdateState>) {
+    const artifact = this.#getArtifact(messageId);
+
+    if (!artifact) {
+      return;
+    }
+
+    this.artifacts.setKey(messageId, { ...artifact, ...state });
+  }
+
+  async addAction(data: ActionCallbackData) {
+    const { messageId } = data;
+    const artifact = this.#getArtifact(messageId);
+
+    if (!artifact) {
+      return;
+    }
+
+    await artifact.runner.addAction(data);
+  }
+
+  async runAction(data: ActionCallbackData) {
+    const { messageId } = data;
+    const artifact = this.#getArtifact(messageId);
+
+    if (!artifact) {
+      unreachable('Artifact not found');
+    }
+
+    await artifact.runner.runAction(data);
+  }
+
+  #getArtifact(messageId: string | undefined): ArtifactState | undefined {
+    if (!messageId) {
+      return undefined;
+    }
+
+    return this.artifacts.get()[messageId];
+  }
+}
