+++ app/lib/stores/documents/DocumentStore.ts
@@ -1,0 +1,139 @@
+import { atom, type WritableAtom } from 'nanostores';
+import type { FilesStore, FileMap } from '../files';
+import type { EditorStore } from '../editor';
+import type { ScrollPosition } from '~/components/editor/codemirror/types';
+
+export class DocumentStore {
+  #filesStore: FilesStore;
+  #editorStore: EditorStore;
+  unsavedFiles: WritableAtom<Set<string>> = import.meta.hot?.data.unsavedFiles ?? atom(new Set<string>());
+  modifiedFiles = new Set<string>();
+
+  constructor(filesStore: FilesStore, editorStore: EditorStore) {
+    this.#filesStore = filesStore;
+    this.#editorStore = editorStore;
+
+    if (import.meta.hot) {
+      import.meta.hot.data.unsavedFiles = this.unsavedFiles;
+    }
+  }
+
+  setDocuments(files: FileMap) {
+    this.#editorStore.setDocuments(files);
+
+    if (this.#filesStore.filesCount > 0 && this.#editorStore.currentDocument.get() === undefined) {
+      // we find the first file and select it
+      for (const [filePath, dirent] of Object.entries(files)) {
+        if (dirent?.type === 'file') {
+          this.setSelectedFile(filePath);
+          break;
+        }
+      }
+    }
+  }
+
+  setCurrentDocumentContent(newContent: string) {
+    const filePath = this.#editorStore.currentDocument.get()?.filePath;
+
+    if (!filePath) {
+      return;
+    }
+
+    const originalContent = this.#filesStore.getFile(filePath)?.content;
+    const unsavedChanges = originalContent !== undefined && originalContent !== newContent;
+
+    this.#editorStore.updateFile(filePath, newContent);
+
+    const currentDocument = this.#editorStore.currentDocument.get();
+
+    if (currentDocument) {
+      const previousUnsavedFiles = this.unsavedFiles.get();
+
+      if (unsavedChanges && previousUnsavedFiles.has(currentDocument.filePath)) {
+        return;
+      }
+
+      const newUnsavedFiles = new Set(previousUnsavedFiles);
+
+      if (unsavedChanges) {
+        newUnsavedFiles.add(currentDocument.filePath);
+      } else {
+        newUnsavedFiles.delete(currentDocument.filePath);
+      }
+
+      this.unsavedFiles.set(newUnsavedFiles);
+    }
+  }
+
+  setCurrentDocumentScrollPosition(position: ScrollPosition) {
+    const editorDocument = this.#editorStore.currentDocument.get();
+
+    if (!editorDocument) {
+      return;
+    }
+
+    const { filePath } = editorDocument;
+    this.#editorStore.updateScrollPosition(filePath, position);
+  }
+
+  setSelectedFile(filePath: string | undefined) {
+    this.#editorStore.setSelectedFile(filePath);
+  }
+
+  async saveFile(filePath: string) {
+    const documents = this.#editorStore.documents.get();
+    const document = documents[filePath];
+
+    if (document === undefined) {
+      return;
+    }
+
+    await this.#filesStore.saveFile(filePath, document.value);
+
+    const newUnsavedFiles = new Set(this.unsavedFiles.get());
+    newUnsavedFiles.delete(filePath);
+
+    this.unsavedFiles.set(newUnsavedFiles);
+  }
+
+  async saveCurrentDocument() {
+    const currentDocument = this.#editorStore.currentDocument.get();
+
+    if (currentDocument === undefined) {
+      return;
+    }
+
+    await this.saveFile(currentDocument.filePath);
+  }
+
+  resetCurrentDocument() {
+    const currentDocument = this.#editorStore.currentDocument.get();
+
+    if (currentDocument === undefined) {
+      return;
+    }
+
+    const { filePath } = currentDocument;
+    const file = this.#filesStore.getFile(filePath);
+
+    if (!file) {
+      return;
+    }
+
+    this.setCurrentDocumentContent(file.content);
+  }
+
+  async saveAllFiles() {
+    for (const filePath of this.unsavedFiles.get()) {
+      await this.saveFile(filePath);
+    }
+  }
+
+  getFileModifcations() {
+    return this.#filesStore.getFileModifications();
+  }
+
+  resetAllFileModifications() {
+    this.#filesStore.resetFileModifications();
+  }
+}
