+++ app/utils/constants.ts
@@ -1,3 +1,5 @@
-export const WORK_DIR_NAME = 'project';
+export const WORK_DIR_NAME = 'workspace';
 export const WORK_DIR = `/home/${WORK_DIR_NAME}`;
 export const MODIFICATIONS_TAG_NAME = 'bolt_file_modifications';
+
+export const WORK_DIR_REGEX = new RegExp(`^${WORK_DIR.split('/').slice(0, -1).join('/').replaceAll('/', '\\/')}/`);
